Solution approach:
Intuition :
So this question is asking to calculate the total number of valid partitions . A partition is only valid if the difference of the sum of left and right subarray is even.

Approach
Initialize count as zero.
SIMPLE MATH APPROACH :
The difference between the two numbers is even only if their parity is same.
Now what is Parity ?
Parity tells us whether the sum is even or odd.
The parity is even if sum % 2 == 0
and The parity is odd if sum % 2 == 1
The concept is to check whether the parity of sums of left and right subarray is equal or not.
If it is equal then we increment the count else we dont.
Also , since in the question none of the subarrays can be empty
so we iteriate till 0 to n -2 only
Example:
arr = {10 , 10, 3,7,6}
First we calculate total sum = 10 + 10 + 3 + 7 + 6 = 36
for i = 0,
now we calculate leftSum = 10
=> rightSum = totalSum - leftSum = 26
let's check parity of both sum
Since 26 % 2 == 0 && 10 % 2 == 0 ..we increment count , count = 1
for i = 1,
now we calculate leftSum = 20
=> rightSum = totalSum - leftSum = 16
let's check parity of both sum
Since 20 % 2 == 0 && 16 % 2 == 0 ..we increment count , count = 2
for i = 2,
now we calculate leftSum = 23
=> rightSum = totalSum - leftSum = 13
let's check parity of both sum
Since 23 % 2 == 1 && 13 % 2 == 1 ..we increment count , count = 3
for i = 3,
now we calculate leftSum = 30
=> rightSum = totalSum - leftSum = 6
let's check parity of both sum
Since 30 % 2 == 0 && 6 % 2 == 0 ..we increment count , count = 4
for i = 4,
now we calculate leftSum = 36
=> rightSum = totalSum - leftSum = 0
// Not possible since right subarray cannot be empty

Output : 4

Complexity
Time complexity:
O(N)
Space complexity:
O(1) since no extra space is used.